// src/profile/dto/enum-horoscope.dto.ts
export enum Horoscope {
  ARIES = 'aries',            
  TAURUS = 'taurus',          
  GEMINI = 'gemini',          
  CANCER = 'cancer',          
  LEO = 'leo',                
  VIRGO = 'virgo',            
  LIBRA = 'libra',            
  SCORPIO = 'scorpio',        
  SAGITTARIUS = 'sagittarius',
  CAPRICORN = 'capricorn',    
  AQUARIUS = 'aquarius',      
  PISCES = 'pisces',          
}